chore(jspm): separation of element of the library for future better testing
