+++ public/app/angular-stomp.js
@@ -1,5 +1,6 @@
 import angular from 'angular';
 import Stomp from 'stompjs';
+import ngstompProvider from './modules/provider';
 
 export default angular
     .module('AngularStompDK', [])
